backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
fpdf2>=2.7.0
